JuiceBot/requirements.txt
discord.py==0.16.7
praw==4.5.0